% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_palette.R
\name{custom_palette_c25}
\alias{custom_palette_c25}
\title{Custom color palette: 25 distinct colors}
\usage{
custom_palette_c25()
}
\value{
a vector of 25 color names
}
\description{
Custom color palette: 25 distinct colors
}
\examples{
custom_palette_c25()
}
