% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_proj}
\alias{demo_proj}
\title{Demographic projection of a bird population over 30 years, under 2 scenarios: sc0 and sc1}
\format{
An object of class \code{array} of dimension 4 x 30 x 2 x 100.
}
\usage{
demo_proj
}
\description{
An 4D array containing population size by age class (dim 1), each year (dim 2),
for both scenarios (dim 3), and for each simulation run (dim 4).
}
\keyword{datasets}
