% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_impact.R
\name{density_impact}
\alias{density_impact}
\title{Plot the density function of the relative impact for each scenario}
\usage{
density_impact(
  N,
  show_CI = 0.95,
  center = "median",
  sel_sc = "all",
  xlims = NULL,
  percent = TRUE,
  xlab = "Relative impact (\%)",
  ylab = "Probability density",
  Legend = NULL,
  legend_position = "right",
  text_size = "large",
  ...
)
}
\arguments{
\item{N}{a 4-D array containing demographic projection outputs}

\item{show_CI}{value between 0 and 1. The quantile to display on the plot}

\item{center}{choose either "median" or "mode" to display central value}

\item{sel_sc}{scenario to display on the plot. Either "all" or the ID number of a given scenario.}

\item{xlims}{a vector of 2 values. x-axis limits (lower and upper).}

\item{percent}{a logical value indicating whether the impact should be displayed in \% (y axis).
If FALSE, the impact value displayed is between 0 and -1 (negative impact).}

\item{xlab}{a character string. Label for the x axis.}

\item{ylab}{a character string. Label for the y axis.}

\item{Legend}{a vector of character strings. The legend to show on the side of the plot.}

\item{legend_position}{Position of the legend : "bottom", "right", etc. See option in ggplot2.}

\item{text_size}{Size of labels and legend text. Either "large" or "small".}

\item{...}{any other graphical input similar to the R plot function}
}
\value{
a plot of the relative impact of each scenario.
}
\description{
Plot the density function of the relative impact for each scenario
}
