% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitation.R
\name{elicitation}
\alias{elicitation}
\title{Function to estimate parameters from expert elicitation}
\usage{
elicitation(vals, Cp, weights = 1, lower = 0, upper = Inf)
}
\arguments{
\item{vals}{a matrix containing the elicited min, best and max value from a set of experts.
One expert per column. ncol = number of expert elicited.}

\item{Cp}{a vector containing the elicited coverage probabilities from the same experts.}

\item{weights}{weights attributed to each expert}

\item{lower}{lower limit}

\item{upper}{upper limit}
}
\value{
a list
}
\description{
Function to estimate parameters from expert elicitation
}
\examples{

## Elicited values for 4 experts
E1 <- c(50, 70, 100)
E2 <- c(200, 240, 280)
E3 <- c(100, 180, 300)
E4 <- c(120, 160, 220)
vals <- matrix(cbind(E1, E2, E3, E4), nrow = 3, byrow = FALSE)

## Elicited coverage probabilities from the same 4 experts
Cp = c(0.80, 0.90, 0.90, 0.70)

## Weights attributed to each expert
weights = c(1, 0.2, 0.2, 0.1)

## Estimation
elicitation(vals, Cp, weights = weights)

}
