% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_DD.R
\name{infer_DD}
\alias{infer_DD}
\title{Infer Densisty-dependence Parameters}
\usage{
infer_DD(
  lambda_MAX = NULL,
  K = NULL,
  theta = 1,
  pop_size_current,
  pop_growth_current
)
}
\arguments{
\item{lambda_MAX}{a strictly positive number. Maximum (theoretical) population growth rate.}

\item{K}{a strictly positive number. Carrying capacity (= maximum size that the population can ever reach)}

\item{theta}{a strictly positive number. Parameter defining the shape of the density-dependence relationship.
The relationship is defined as : lambda <- 1 + rMAX*(1-(N/K)^theta)
where rMAX = lambda_MAX - 1}

\item{pop_size_current}{a strictly positive number. Current population size.}

\item{pop_growth_current}{a strictly positive number. Current population growth rate.}
}
\value{
a list of length 2, containing K and lambda_MAX.
}
\description{
infer_DD calculates the missing value of a parameter (either K or rMAX, whichever is NULL) in a density-dependence relationship.
}
\examples{
## Infer K
infer_DD(lambda_MAX = 1.15, K = NULL, theta = 1, pop_size_current = 200, pop_growth_current = 1.08)
## Infer lambda_MAX
infer_DD(lambda_MAX = NULL, K = 2000, theta = 1, pop_size_current = 200, pop_growth_current = 1.08)


}
