% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_rMAX.R
\name{infer_rMAX}
\alias{infer_rMAX}
\title{Infer the rMAX}
\usage{
infer_rMAX(
  K,
  theta = 1,
  pop_size_current,
  pop_growth_current,
  rMAX_theoretical = Inf
)
}
\arguments{
\item{K}{a strictly positive number. Carrying capacity (= maximum size that the population can ever reach)}

\item{theta}{a strictly positive number. Parameter defining the shape of the density-dependence relationship.
The relationship is defined as : r <- rMAX*(1-(N/K)^theta)
Note lambda = r + 1}

\item{pop_size_current}{a strictly positive number. Current population size.}

\item{pop_growth_current}{a strictly positive number. Current population growth rate.}

\item{rMAX_theoretical}{the maximum value for rMAX, usually calculated from the function rMAX_spp. See ?rMAX_spp for details.}
}
\value{
the r_MAX value.
}
\description{
Calculates the rMAX of a density-dependence relationship, based on K and current population status (size and trend)
}
\examples{
## Infer rMAX
infer_rMAX(K = 2000, theta = 1, pop_size_current = 200, pop_growth_current = 1.08)


}
