% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_theta.R
\name{infer_theta}
\alias{infer_theta}
\title{Infer the rMAX}
\usage{
infer_theta(K, pop_size_current, pop_growth_current, rMAX)
}
\arguments{
\item{K}{a strictly positive number. Carrying capacity (= maximum size that the population can ever reach)}

\item{pop_size_current}{a strictly positive number. Current population size.}

\item{pop_growth_current}{a strictly positive number. Current population growth rate.}

\item{rMAX}{the maximum rate of population growth.}
}
\value{
the theta value of a theta-logistic population growth model.
}
\description{
Calculates the rMAX of a density-dependence relationship, based on K and current population status (size and trend)
}
