% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params.R
\name{init_calib}
\alias{init_calib}
\title{Set initial values for the calibration of vital rates.
This function can be used  to speed up the optim process of the 'calibrate_params' function.}
\usage{
init_calib(s, f, lam0)
}
\arguments{
\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age clas}

\item{lam0}{the desired population growth rate - the one to be matched}
}
\value{
a vetor of initial values for the calibration of survival and fecundity values
}
\description{
Set initial values for the calibration of vital rates.
This function can be used  to speed up the optim process of the 'calibrate_params' function.
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)
init_calib(s = s, f = f, lam0 = 1.08)

}
