% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{M3_WithDD_noDemoStoch}
\alias{M3_WithDD_noDemoStoch}
\title{Demographic model with Density-Dependence, but without Demographic Stochasticity
In addition to natural survivals and fecundities, this demographic model includes
a specific harvest / fatality parameter.
NOTE : This is a POST-BREEDING demographic model.}
\usage{
M3_WithDD_noDemoStoch(
  N1,
  s,
  f,
  h,
  DD_params,
  use_ref_vr = FALSE,
  s_corr_factor = NULL,
  f_corr_factor = NULL
)
}
\arguments{
\item{N1}{a vector of population sizes for each age class at time t1}

\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age class}

\item{h}{a number. The harvest or fatality rate}

\item{DD_params}{a list containing the 3 parameters required to model density-dependence :
rMAX (maximum population intinsic rate of increase: lambda_max - 1),
K (carrying capacity), and
theta (shape of DD relationshp)}

\item{use_ref_vr}{Not used in this model, as there is no demographic stochasticity.}

\item{s_corr_factor}{Not used in this model, as there is no demographic stochasticity.}

\item{f_corr_factor}{Not used in this model, as there is no demographic stochasticity.}
}
\value{
a vector of population sizes for each age class at time t2
}
\description{
Demographic model with Density-Dependence, but without Demographic Stochasticity
In addition to natural survivals and fecundities, this demographic model includes
a specific harvest / fatality parameter.
NOTE : This is a POST-BREEDING demographic model.
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)
N1 <- c(50, 60, 75, 100)
h <- 0.05
DD_params <- list(rMAX = 0.15, K = 1200, theta = 1)
M3_WithDD_noDemoStoch(N1, s, f, h,DD_params = DD_params)

}
