% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{M4_WithDD_WithDemoStoch}
\alias{M4_WithDD_WithDemoStoch}
\title{Demographic model with Density-Dependence and INCLUDING Demographic Stochasticity
In addition to natural survivals and fecundities, this demographic model includes
a specific harvest / fatality parameter.
NOTE : This is a POST-BREEDING demographic model.}
\usage{
M4_WithDD_WithDemoStoch(N1, s, f, h, DD_params)
}
\arguments{
\item{N1}{a vector of population sizes for each age class at time t1}

\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age class}

\item{h}{a number. The harvest or fatality rate}

\item{DD_params}{density-dependence parameters. Not used in this model.}
}
\value{
a vector of population sizes for each age class at time t2
}
\description{
Demographic model with Density-Dependence and INCLUDING Demographic Stochasticity
In addition to natural survivals and fecundities, this demographic model includes
a specific harvest / fatality parameter.
NOTE : This is a POST-BREEDING demographic model.
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)
N1 <- c(50, 60, 75, 100)
h <- 0.05
DD_params <- list(rMAX = 0.15, K = 1200, theta = 1)
M4_WithDD_WithDemoStoch(N1, s, f, h, DD_params = DD_params)

}
