% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj.R
\name{make_transparent}
\alias{make_transparent}
\title{Function to control color transparency on a plot}
\usage{
make_transparent(someColor, percent = 100)
}
\arguments{
\item{someColor}{some color name, such as "green", "red", etc.}

\item{percent}{the desired \% of transparency, between 0 (no transparency) and 100 (full transparency) .}
}
\value{
a Hex Code including the desired transparency level, such as #RRGGBB7F
}
\description{
Function to control color transparency on a plot
}
\examples{
make_transparent("green", percent=50)
}
