% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params.R
\name{match_lam_delta}
\alias{match_lam_delta}
\title{Function to calculate the difference to apply to each vital rates to match the desired
population growth rate (lambda).
This function is used in the 'init_calib' function.}
\usage{
match_lam_delta(diff_rel_lam, s, f)
}
\arguments{
\item{diff_rel_lam}{a number. The relative difference of lambda (population growth rate)
to be matched}

\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age clas}
}
\value{
a vector of (absolute) difference to apply
}
\description{
Function to calculate the difference to apply to each vital rates to match the desired
population growth rate (lambda).
This function is used in the 'init_calib' function.
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)

# Match for a 5\% decrease in lambda
match_lam_delta(diff_rel_lam = -0.05, s, f)
}
