% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitation.R
\name{plot_elicitation}
\alias{plot_elicitation}
\title{Function to plot the expert elicitation curves}
\usage{
plot_elicitation(out)
}
\arguments{
\item{out}{the output list from the elicitation function}
}
\value{
a plot of the estimated distirbution (elitation curves) of elicited parameters
}
\description{
Function to plot the expert elicitation curves
}
\examples{
## Elicited values for 4 experts
E1 <- c(50, 70, 100)
E2 <- c(200, 240, 280)
E3 <- c(100, 180, 300)
E4 <- c(120, 160, 220)
vals <- matrix(cbind(E1, E2, E3, E4), nrow = 3, byrow = FALSE)

## Elicited coverage probabilities from the same 4 experts
Cp = c(0.80, 0.90, 0.90, 0.70)

## Estimate
out <- elicitation(vals, Cp)

## Plot it
plot_elicitation(out)

}
