% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_vector.R
\name{pop_vector}
\alias{pop_vector}
\title{Build a vector}
\usage{
pop_vector(pop_size, pop_size_type = "Npair", s, f)
}
\arguments{
\item{pop_size}{a single number. Total Population Size or Number of Pairs.}

\item{pop_size_type}{character value indicating if the provided value pop_size correpsonds to Total Population Size ("Ntotal")
or the Number of Pairs ("Npair"). A stable age distribution is used to infer the size of each age class.}

\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age class}
}
\value{
a vector of sizes for each age class. The number of age class is deduced from the size of the 's' vector.
}
\description{
Build a vector
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)
pop_vector(pop_size = 200, pop_size_type = "Npair", s, f)

}
