% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles_impact.R
\name{quantiles_impact}
\alias{quantiles_impact}
\title{Calculate quantile values of impact at time horizon}
\usage{
quantiles_impact(dr_N, show_quantile = 0.975, show_CI = 0.95, percent = TRUE)
}
\arguments{
\item{dr_N}{a 3-D array (year, nsim, nsc) containing the output : relative difference of population size}

\item{show_quantile}{value between 0 and 1. The quantile to of the CDF to calculate}

\item{show_CI}{value between 0 and 1. The limits of C.I. to calculate}

\item{percent}{a logical value indicating whether the impact should be displayed in \% (y axis).
If FALSE, the impact value displayed is between 0 and -1 (negative impact).}
}
\value{
the CI and quantiles values
}
\description{
Calculate quantile values of impact at time horizon
}
