% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAX_spp.R
\name{rMAX_spp}
\alias{rMAX_spp}
\title{Rmax value for a given species}
\usage{
rMAX_spp(surv, afr)
}
\arguments{
\item{surv}{Adult survival. Single value between 0 and 1.}

\item{afr}{Age at first reproduction. Single integer value.}
}
\value{
the theoretical Rmax of the species
}
\description{
Calculate the theoretical Rmax of a species using the equations of Niel and Lebreton 2005.
There are two different equation depending on whether the species is longlived or "short-lived".
Here, we consider a species to be "shortlived" if age at first reproduction <= 2 and adult survival <= 0.8.

References.
Niel, C., and J. Lebreton. 2005. Using demographic invariants to detect overharvested bird
populations from incomplete data. Conservation Biology 19:826-835.
}
\examples{
rMAX_spp(surv = 0.6, afr = 2)

}
