% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{run0}
\alias{run0}
\title{Demographic projection under 5 scenarios}
\format{
An object of class \code{list} of length 3.
}
\usage{
run0
}
\description{
A list Also a 4D array containing results : population size by age class (dim 1), each year (dim 2),
for all scenarios (dim 3), and for each simulation run (dim 4).
}
\keyword{datasets}
