% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simul.R
\name{run_simul}
\alias{run_simul}
\title{Run multiple population projections (simulations)}
\usage{
run_simul(
  nsim,
  cumuated_impacts,
  fatalities_mean,
  fatalities_se,
  onset_time,
  pop_size_mean,
  pop_size_se,
  pop_size_type,
  pop_growth_mean,
  pop_growth_se,
  survivals,
  fecundities,
  model_demo,
  time_horzion,
  coeff_var_environ,
  fatal_constant
)
}
\arguments{
\item{nsim}{number of simulation}

\item{cumuated_impacts}{Logical. If TRUE, we used the projection model for cumulated impacts.}

\item{fatalities_mean}{a vector (numeric). Average number of fatalities, for each scenario.}

\item{fatalities_se}{a vector (numeric). Standard Error for the number of fatalities, for each scenario (= uncertainties around the values provided).}

\item{onset_time}{a vector (numeric). The times at which each wind farm fatality starts applying.}

\item{pop_size_mean}{a single number. Average population size (either total population or number of pairs - see Ntype below).}

\item{pop_size_se}{Standard Error for population size (= uncertainty around the value provided).}

\item{pop_size_type}{character value indicating if the provided value pop_size correpsonds to Total Population Size ("Ntotal")
or the Number of Pairs ("Npair"). A stable age distribution is used to infer the size of each age class.}

\item{pop_growth_mean}{a number. Average population growth rate (lambda).}

\item{pop_growth_se}{Standard Error for population growth rate (= uncertainty around the value provided).}

\item{survivals}{a vector. Average survival probabilities for each age class.}

\item{fecundities}{a vector of fecundity values for each age class.}

\item{model_demo}{an R object corresponding to the demographic model to be used. The 4 possible models currently are:
M1_noDD_noDemoStoch, M2_noDD_WithDemoStoch, M3_WithDD_noDemoStoch, M4_WithDD_WithDemoStoch,}

\item{time_horzion}{a number. The number of years (time horizon) over which to project the population dynamics.}

\item{coeff_var_environ}{a number. The coefficient of variation to model environment stochasticity.}

\item{fatal_constant}{text (character). Either "h" or "M". Using "h" sets the fatality RATE as the constant value across years.
Using "M" sets the NUMBER of fatalities as the constant value across years.}
}
\value{
a 4D array containing the size of each age class (dim 1), for each year (dim 2), each scenario (dim 3), and
each simulation iteration (dim 4)
}
\description{
Run multiple population projections (simulations)
}
\examples{
fatalities_mean = c(0, 5, 10, 15, 20)
fatalities_se = fatalities_mean*0.05

pop_size_mean = 200
pop_size_se = 30
pop_size_type = "Npair"

pop_growth_mean = 1
pop_growth_se = 0.03

survivals <- c(0.5, 0.7, 0.8, 0.95)

fecundities <- c(0, 0, 0.05, 0.55)

model_demo = M2_noDD_WithDemoStoch

time_horzion = 30
coeff_var_environ = 0.10
fatal_constant = "h"

run_simul(nsim = 10, cumuated_impacts = FALSE,
           fatalities_mean, fatalities_se, onset_time = NULL,
           pop_size_mean, pop_size_se, pop_size_type,
           pop_growth_mean, pop_growth_se,
           survivals, fecundities,
           model_demo, time_horzion, coeff_var_environ, fatal_constant)


}
