% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gamma.R
\name{sample_gamma_v1}
\alias{sample_gamma_v1}
\title{Draw values from a Custom Gamma Distribution using lower and upper limits}
\usage{
sample_gamma_v1(n, lower, upper, coverage = 0.95)
}
\arguments{
\item{n}{number of value to draw.}

\item{lower}{lower limit}

\item{upper}{upper limit}
}
\value{
a vector of drawn values.
}
\description{
Draw values from a Custom Gamma Distribution using lower and upper limits
}
\examples{
sample_gamma_v1(n = 50, lower = 10, upper = 30)

}
