% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAX_spp.R
\name{theta_spp}
\alias{theta_spp}
\title{Theta value for a given species}
\usage{
theta_spp(rMAX_species)
}
\arguments{
\item{rMAX_species}{Species rMAX value (Between 0 and +Inf)}
}
\value{
the theoretical Rmax of the species
}
\description{
Calculate the theoretical value of the theta parameter (= shape of the Density-dependence relationship) of a species using the equations of Johnson et al. 2012.
The equation is log(theta) = 1.129 - 1.824*rMAX

References.
Johnson, F. A., M. A. H. Walters, and G. S. Boomer. 2012. Allowable levels of take for the trade in
Nearctic songbirds. Ecological Applications 22:1114-1130.
}
\examples{
rMAX_spp(surv = 0.6, afr = 2)

}
