% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Leslie.R
\name{elements_Leslie}
\alias{elements_Leslie}
\title{Extract Leslie matrix elements}
\usage{
elements_Leslie(s, f)
}
\arguments{
\item{s}{a vector of survival probabilities for each age class}

\item{f}{a vector of fecundity values for each age class}
}
\description{
Extract Leslie matrix elements
}
\examples{
s <- c(0.5, 0.7, 0.8, 0.95)
f <- c(0, 0, 0.05, 0.55)
elements_Leslie(s=s, f=f)

}
