% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Calculate Impact Metric
Relative Difference of Population Size DR_N}
\usage{
get_metrics(N)
}
\arguments{
\item{N}{a 4-D array containing demographic projection outputs}
}
\value{
a list of metric outputs : mean, SD, 95\% C.I. of the
}
\description{
Calculate Impact Metric
Relative Difference of Population Size DR_N
}
\examples{
## Import Data
data("demo_proj")

## Calculate the metric
get_metrics(demo_proj)

}
