% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_DD.R
\name{infer_DD}
\alias{infer_DD}
\title{Infer Densisty-dependence Parameters}
\usage{
infer_DD(
  rMAX = NULL,
  K = NULL,
  theta = 1,
  pop_size_current,
  pop_growth_current
)
}
\arguments{
\item{rMAX}{a number. Maximum (theoretical) population intrinsic rate of increase.}

\item{K}{a strictly positive number. Carrying capacity (= maximum size that the population can ever reach)}

\item{theta}{a strictly positive number. Parameter defining the shape of the density-dependence relationship.
The relationship is defined as : r <- rMAX*(1-(N/K)^theta)
Note lambda = r + 1}

\item{pop_size_current}{a strictly positive number. Current population size.}

\item{pop_growth_current}{a strictly positive number. Current population growth rate.}
}
\value{
a list of length 2, containing K and lambda_MAX.
}
\description{
infer_DD calculates the missing value of a parameter (either K or rMAX, whichever is NULL) in a density-dependence relationship.
}
\examples{
## Infer K
infer_DD(rMAX = 0.15, K = NULL, theta = 1, pop_size_current = 200, pop_growth_current = 1.08)
## Infer rMAX
infer_DD(rMAX = NULL, K = 2000, theta = 1, pop_size_current = 200, pop_growth_current = 1.08)


}
