% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact.R
\name{plot_impact}
\alias{plot_impact}
\title{Plot demographic trajectories}
\usage{
plot_impact(N, ...)
}
\arguments{
\item{N}{a 4-D array containing demographic projection outputs}

\item{...}{any other graphical input similar to the R plot function}
}
\value{
a plot of the relative impact of each scenario.
}
\description{
Plot demographic trajectories
}
\examples{
# plot_impact(demo_proj, xlab = "year", ylab = "pop size")

}
