% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact.R
\name{plot_impact}
\alias{plot_impact}
\title{Plot the relative impact for each scenario}
\usage{
plot_impact(
  N,
  onset_year = NULL,
  percent = TRUE,
  xlab = "Year",
  ylab = "Relative impact (\%)",
  Legend = NULL,
  ...
)
}
\arguments{
\item{N}{a 4-D array containing demographic projection outputs}

\item{onset_year}{a vector containing the years of each wind farm start being active
(thus, the year at whihc each fatality value starts kicking in)}

\item{percent}{a logical value indicating whether the impact should be displayed in \% (y axis).
If FALSE, the impact value displayed is between 0 and -1 (negative impact).}

\item{xlab}{a character string. Label for the x axis.}

\item{ylab}{a character string. Label for the y axis.}

\item{Legend}{a vector of character strings. The legend to show on the side of the plot.}

\item{...}{any other graphical input similar to the R plot function}
}
\value{
a plot of the relative impact of each scenario.
}
\description{
Plot the relative impact for each scenario
}
\examples{


}
