% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot demographic trajectories}
\usage{
plot_traj(N, len = 50, ...)
}
\arguments{
\item{N}{a 4-D array containing demographic projection outputs}

\item{len}{number of individual trajectories to show on the plot}

\item{...}{any other graphical input similar to the R plot function}
}
\value{
a plot of individual trajectories, from each simulation run,
and the average trajectory of each scenario.
Black = reference scenario (sc0). Red = fatality scenario (sc1)
}
\description{
Plot demographic trajectories
}
\examples{
plot_traj(demo_proj, xlab = "year", ylab = "pop size")
}
