% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAX_spp.R
\name{rMAX_spp}
\alias{rMAX_spp}
\title{Rmax of a species}
\usage{
rMAX_spp(surv, afr)
}
\arguments{
\item{surv}{Adult survival. Single value between 0 and 1.}

\item{afr}{Age at first reproduction. Single integer value.}
}
\value{
the theoretical Rmax of the species
}
\description{
Calculate the theoretical Rmax of a species using the equation of Niel and Lebreton 2005.

References.

Niel, C., and J. Lebreton. 2005. Using demographic invariants to detect overharvested bird
populations from incomplete data. Conservation Biology 19:826–835.
}
\examples{
rMAX_spp(surv = 0.6, afr = 2)

}
