% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gamma.R
\name{sample_gamma_v1}
\alias{sample_gamma_v1}
\title{Draw values from a Custom Gamma Distribution using mean and standard deviation as inputs}
\usage{
sample_gamma_v1(n, mu, sd, r = 3)
}
\arguments{
\item{n}{number of value to draw.}

\item{mu}{mean}

\item{sd}{standard deviation}

\item{r}{scaling factor. Used to ensure a Normal shape for small mean values.}
}
\value{
a vector of drawn values.
}
\description{
Draw values from a Custom Gamma Distribution using mean and standard deviation as inputs
}
\examples{
sample_gamma_v1(n= 5, mu = 4, sd = 1.5)

}
